Fold pagination regression tests into query-generation test file
Move the pagination tests from a standalone file into
test_query_generation_multiple_dimensions.py. The module-level django_db
marker becomes per-class markers on the existing classes so the new
pagination classes stay DB-free (they only compile SQL).
